Add the config option to determine whether to push the artifacts or not
This PR adds a config option "publish_stage" in the configuration file
to determine whether this commit will push new artifacts into the
staging URL or not. By default, it is set to false, meaning it will
not push new artifacts. We can set it to true in order to push new
artifacts.