Add script to package the documents
This PR adds another flag called UPDATE_DOC to determine if the commit
is going to change the documents or not.

Besides, it re-arranged the logic when a PR is merged into master.
It separated the conditions based on the TRAVIS_EVENT_TYPE, so that
pull requests and pushes can be more trackable.